Added calculation and sending of temperature fractional part.
Temperature from SI7005 is now represented as uint32
containing two complete and two fractional digits.

+++ esp8266-kinetis/Sources/si7005/si7005.h
@@ -20,6 +20,6 @@ typedef enum {
 
 si7005_status_code_e si7005_open(LDD_TDeviceData **si7005_device_data_p);
 uint8_t read_rH(uint8_t *rH);
-uint8_t read_temperature(uint8_t *temperature);
+uint8_t read_temperature(uint32_t *temperature);
 
 #endif /* SI7005_H_ */
